Moved css to less file and removed all css-in-js, created directory signup-confirmation in components
