+++ src/components/signup-confirmation/SignupConfirmation.less
@@ -1,0 +1,48 @@
+.input-style {
+  borderRadius: 5px;
+  height: 48px;
+  backgroundColor: #E5E5E5;
+  marginBottom: 15px;
+  marginTop: 6px;
+  fontSize: 18px;
+}
+
+.radio-style {
+  display: block;
+  height: 30px;
+  lineHeight: 30px;
+}
+
+.submit-button {
+  display: inline-block;
+  margin: 0 8px;
+  width: 195px;
+  height: 45px;
+  borderRadius: 5px;
+  background: #505050;
+}
+
+.back-button {
+  display: inline-block;
+  margin: 0 8px;
+  width: 195px;
+  height: 45px;
+  borderRadius: 5px;
+  background: #C5C5C5;
+}
+
+.photo-release-div {
+  margin-top: 36px;
+  fontSize: 18px;
+}
+
+.checkbox-div {
+  marginTop: 36px;
+  fontSize: 18px;
+}
+
+.input-item {
+  display: inline-block;
+  width: calc(50% - 8px);
+  margin: 0 8px 0px 0px;
+}
